+++ src/ui/util.rs
@@ -183,7 +183,7 @@ mod test {
     #[test]
     fn test_container_height_username_padding_zero() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options().parse(&["--container-padding", "0"]).ok();
+        greeter.container_padding = 1;
         greeter.mode = Mode::Username;
 
         assert_eq!(get_height(&greeter), 3);
@@ -197,7 +197,7 @@ mod test {
     #[test]
     fn test_container_height_username_padding_one() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options().parse(&["--container-padding", "1"]).ok();
+        greeter.container_padding = 2;
         greeter.mode = Mode::Username;
 
         assert_eq!(get_height(&greeter), 5);
@@ -213,7 +213,7 @@ mod test {
     #[test]
     fn test_container_height_username_greeting_padding_one() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options().parse(&["--container-padding", "1"]).ok();
+        greeter.container_padding = 2;
         greeter.greeting = Some("Hello".into());
         greeter.mode = Mode::Username;
 
@@ -232,7 +232,7 @@ mod test {
     #[test]
     fn test_container_height_password_greeting_padding_one_prompt_padding_1() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options().parse(&["--container-padding", "1"]).ok();
+        greeter.container_padding = 2;
         greeter.greeting = Some("Hello".into());
         greeter.mode = Mode::Password;
         greeter.prompt = Some("Password:".into());
@@ -251,9 +251,8 @@ mod test {
     #[test]
     fn test_container_height_password_greeting_padding_one_prompt_padding_0() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options()
-            .parse(&["--container-padding", "1", "--prompt-padding", "0"])
-            .ok();
+        greeter.container_padding = 2;
+        greeter.prompt_padding = 0;
         greeter.greeting = Some("Hello".into());
         greeter.mode = Mode::Password;
         greeter.prompt = Some("Password:".into());
@@ -264,7 +263,7 @@ mod test {
     #[test]
     fn test_rect_bounds() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options().parse(&["--width", "50"]).ok();
+        greeter.width = 50;
 
         let (x, y, width, height) = get_rect_bounds(&greeter, Rect::new(0, 0, 100, 100), 1);
 
@@ -281,9 +280,8 @@ mod test {
     #[test]
     fn input_width() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options()
-            .parse(&["--width", "40", "--container-padding", "1"])
-            .ok();
+        greeter.width = 40;
+        greeter.container_padding = 2;
 
         let input_width = get_input_width(&greeter, 40, &Some("Username:".into()));
 
@@ -293,9 +291,8 @@ mod test {
     #[test]
     fn greeting_height_one_line() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options()
-            .parse(&["--width", "15", "--container-padding", "1"])
-            .ok();
+        greeter.width = 15;
+        greeter.container_padding = 2;
         greeter.greeting = Some("Hello World".into());
 
         let (_, height) = get_greeting_height(&greeter, 1, 0);
@@ -306,9 +303,8 @@ mod test {
     #[test]
     fn greeting_height_two_lines() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options()
-            .parse(&["--width", "8", "--container-padding", "1"])
-            .ok();
+        greeter.width = 8;
+        greeter.container_padding = 2;
         greeter.greeting = Some("Hello World".into());
 
         let (_, height) = get_greeting_height(&greeter, 1, 0);
@@ -319,9 +315,8 @@ mod test {
     #[test]
     fn ansi_greeting_height_one_line() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options()
-            .parse(&["--width", "15", "--container-padding", "1"])
-            .ok();
+        greeter.width = 15;
+        greeter.container_padding = 2;
         greeter.greeting = Some("\x1b[31mHello\x1b[0m World".into());
 
         let (text, height) = get_greeting_height(&greeter, 1, 0);
@@ -339,9 +334,8 @@ mod test {
     #[test]
     fn ansi_greeting_height_two_lines() {
         let mut greeter = Greeter::default();
-        greeter.config = Greeter::options()
-            .parse(&["--width", "8", "--container-padding", "1"])
-            .ok();
+        greeter.width = 8;
+        greeter.container_padding = 2;
         greeter.greeting = Some("\x1b[31mHello\x1b[0m World".into());
 
         let (text, height) = get_greeting_height(&greeter, 1, 0);